Removed limit
The limit of $geoNear would cause the city name input to be ignored if the city entered wasn't within the limit. 
For example, if a user misentered the latitude coordinate by an error of 5, the geonear aggregate would never contain the input city (unless there was another city named the same at those coordinates, which works out fine).