Add a suggestion engine
It comes with scoring capabilities, with built-in scorers for match
length, population, and distance from a coordinates hint.

It's also capable to store the suggestions in different data structures,
with a built-in trie for efficient matching.

A default engine configuration is supplied in 'src/suggestions.js',
using the applicaton data.