Implemented loading the cities list.
Using the csv-parser package (https://github.com/wdavidw/node-csv-parse) I load and parse the TSV list of the cities on the script load. I made it syncronous to make sure that the list is available _before_ the server is initialized (and they're kept in the memory).

It might be not the fastest and/or best way, but will work for now.