Calculate score based on the name match
Created the simple algorithm to calculate the match score based on the name match. If exact match, the score is 100%, if not, calculated based on position and the number of symbols matched. If the position of the search query isn't in the beginning of the matched city, the score is squared => goes down in the list.