Included query length calculation to score and sort suggestion results
To account for when latitude and longitude is not provided, I've include a very basic calculation that checks query length against the length of the city. This is a very basic approach to account for the query given. A better approach would be to calculate the score based on total number of cities in the database and the frequency that the word is searched by users.

Other changes in this document include:
- 100% match returns only one value
- Fixed variable name from country to city
- Fixed 404 status for when no city is found